Added new component Vote and added more to the cards
Added new component Vote.vue that takes 3 props (isLike, isButton, isActive) and emits  (@clicked), it serves as a reusable component for the hands up and hands down icon, and to function either as a button or an icon.

Added most of the finishing touches to the list view of the cards.

Added the vote div buttons and their respective funcionality (With the localstorage as a method of saving the information)

End of the first day