Set the grid card with all the missing info
Filled the card component with the grid information and their respective styles, added a new color for the hover to be noticed on grid (black) and added a trimmedName for the names of the people. Also added the first decimal to the percentage of the likes/dislikes, also removed the word-break in the description.